Move test for TableSource to testComponentInterface.
